refactor(theme): split widget factory into theme_widgets
theme.c was colours, fonts, sizing and an LVGL widget factory in one
file. Move the widget builders/stylers to theme_widgets.{c,h} and the
private COLOR_* macros to theme_palette.h, leaving theme.c/.h as the
design tokens. Widget consumers now include theme_widgets.h; token-only
files depend on theme.h alone.
