+++ main/pages/capture_entropy.c
@@ -14,7 +14,7 @@
 
 #include "../components/video/video.h"
 #include "../ui/dialog.h"
-#include "../ui/theme.h"
+#include "../ui/theme_widgets.h"
 #include "../utils/memory_utils.h"
 #include "../utils/secure_mem.h"
