+++ main/pages/home/backup/backup_menu.c
@@ -4,7 +4,7 @@
 #include "../../../core/storage.h"
 #include "../../../ui/dialog.h"
 #include "../../../ui/menu.h"
-#include "../../../ui/theme.h"
+#include "../../../ui/theme_widgets.h"
 #include "../../store_mnemonic.h"
 #include "mnemonic_qr.h"
 #include "mnemonic_words.h"
