+++ main/pages/new_mnemonic/entropy_from_camera.c
@@ -3,7 +3,7 @@
 #include "entropy_from_camera.h"
 #include "../../ui/dialog.h"
 #include "../../ui/input_helpers.h"
-#include "../../ui/theme.h"
+#include "../../ui/theme_widgets.h"
 #include "../../ui/word_selector.h"
 #include "../capture_entropy.h"
 #include <lvgl.h>
