+++ main/pages/passphrase.c
@@ -1,7 +1,7 @@
 #include "passphrase.h"
 #include "../ui/dialog.h"
 #include "../ui/input_helpers.h"
-#include "../ui/theme.h"
+#include "../ui/theme_widgets.h"
 #include <lvgl.h>
 #include <stdio.h>
