+++ main/pages/pin/pin_page.c
@@ -8,7 +8,7 @@
 #include "../../core/pin.h"
 #include "../../ui/dialog.h"
 #include "../../ui/input_helpers.h"
-#include "../../ui/theme.h"
+#include "../../ui/theme_widgets.h"
 #include "../../utils/secure_mem.h"
 
 #include <esp_system.h>
