+++ main/pages/pin/pin_settings.c
@@ -5,7 +5,7 @@
 #include "../../ui/dialog.h"
 #include "../../ui/input_helpers.h"
 #include "../../ui/menu.h"
-#include "../../ui/theme.h"
+#include "../../ui/theme_widgets.h"
 #include "../../utils/session.h"
 #include "pin_page.h"
