+++ main/pages/screensaver.c
@@ -1,6 +1,6 @@
 #include "screensaver.h"
 #include "ui/assets/kern_logo_lvgl.h"
-#include "ui/theme.h"
+#include "ui/theme_widgets.h"
 #include <esp_random.h>
 
 static lv_obj_t *scr_container;
