+++ main/pages/shared/kef_decrypt_page.c
@@ -12,7 +12,7 @@
 #include "../../core/kef.h"
 #include "../../ui/dialog.h"
 #include "../../ui/input_helpers.h"
-#include "../../ui/theme.h"
+#include "../../ui/theme_widgets.h"
 #include "../../utils/secure_mem.h"
 #include <esp_task_wdt.h>
 #include <freertos/FreeRTOS.h>
