+++ main/pages/store_descriptor.c
@@ -7,7 +7,7 @@
 #include "../core/wallet.h"
 #include "../ui/dialog.h"
 #include "../ui/input_helpers.h"
-#include "../ui/theme.h"
+#include "../ui/theme_widgets.h"
 #include "shared/kef_encrypt_page.h"
 
 #include <lvgl.h>
