+++ main/qr/scanner.c
@@ -5,7 +5,7 @@
 #include "../core/settings.h"
 #include "../ui/dialog.h"
 #include "../ui/input_helpers.h"
-#include "../ui/theme.h"
+#include "../ui/theme_widgets.h"
 #include "../utils/memory_utils.h"
 #include "parser.h"
 #include <bsp/esp-bsp.h>
