+++ main/ui/theme.h
@@ -47,30 +47,6 @@ int theme_corner_button_height(void);
 int theme_small_padding(void);
 int theme_logo_size(void);
 
-void theme_apply_screen(lv_obj_t *obj);
-lv_obj_t *theme_create_page_container(lv_obj_t *parent);
-void theme_apply_frame(lv_obj_t *frame);
-void theme_apply_solid_rectangle(lv_obj_t *target_rectangle);
-void theme_apply_label(lv_obj_t *label, bool is_secondary);
-void theme_apply_button_label(lv_obj_t *label, bool is_secondary);
-void theme_apply_touch_button(lv_obj_t *btn, bool is_primary);
-void theme_apply_btnmatrix(lv_obj_t *btnmatrix);
-lv_obj_t *theme_create_button(lv_obj_t *parent, const char *text,
-                              bool is_primary);
-lv_obj_t *theme_create_label(lv_obj_t *parent, const char *text,
-                             bool is_secondary);
-lv_obj_t *theme_create_page_title(lv_obj_t *parent, const char *text);
-void theme_apply_transparent_container(lv_obj_t *obj);
-lv_obj_t *theme_create_scroll_column(lv_obj_t *parent, int32_t pad,
-                                     int32_t gap);
-lv_obj_t *theme_create_separator(lv_obj_t *parent, lv_color_t color);
-lv_obj_t *theme_create_button_row(lv_obj_t *parent, int32_t gap);
-lv_obj_t *theme_create_flex_row(lv_obj_t *parent);
-lv_obj_t *theme_create_flex_column(lv_obj_t *parent);
-lv_obj_t *theme_create_dropdown(lv_obj_t *parent, const char *options);
-// Square white-background container for hosting an lv_qrcode. `inner_pad` is
-// the QR's white quiet zone (in pixels).
-lv_obj_t *theme_create_qr_container(lv_obj_t *parent, int32_t size,
-                                    int32_t inner_pad);
+// Widget builders (theme_create_*/theme_apply_*) live in theme_widgets.h.
 
 #endif // THEME_H
