+++ simulator/CMakeLists.txt
@@ -130,6 +130,7 @@ set(SIM_SOURCES
     ${CMAKE_CURRENT_SOURCE_DIR}/platform/esp_idf_stubs/stubs.c
     ${APP_UI_DIR}/font_policy.c
     ${APP_UI_DIR}/theme.c
+    ${APP_UI_DIR}/theme_widgets.c
     ${APP_UI_DIR}/assets/kern_logo_lvgl.c
     ${APP_UI_DIR}/assets/kern_logo_font.c
     ${APP_UI_DIR}/assets/icons_16.c
